+++ pkgs/tools/nix/nixos-render-docs/src/nixos_render_docs/docbook.py
@@ -6,19 +6,6 @@ from xml.sax.saxutils import escape, quoteattr
 
 from .md import Renderer
 
-_xml_id_translate_table = {
-    ord('*'): ord('_'),
-    ord('<'): ord('_'),
-    ord(' '): ord('_'),
-    ord('>'): ord('_'),
-    ord('['): ord('_'),
-    ord(']'): ord('_'),
-    ord(':'): ord('_'),
-    ord('"'): ord('_'),
-}
-def make_xml_id(s: str) -> str:
-    return s.translate(_xml_id_translate_table)
-
 class Deflist:
     has_dd = False
